fix: ensure float precision for gate values
Cast gate and cosine similarity tensors to float before detachment 
and CPU conversion to prevent potential type mismatch errors during 
logging and processing.

+++ src/prototype_stream.py
@@ -222,7 +222,7 @@ class SemanticShiftFusion(nn.Module):
 
         s_feat = self.state_emb(state)
         g = self.gate_net(torch.cat([h_ctx, h_proto, s_feat], dim=-1))  # (B, 1)
-        self.last_g = g.detach()
+        self.last_g = g.detach().float()
         alpha = 0.5  # Max context attenuation
         beta = 1.0   # Max prototype shift amplification
         ctx_scaled = (1.0 - alpha * g) * h_ctx
